requirements.txt
Pillow == 9.0.1
matplotlib == 3.5.1
numpy == 1.21.5
opencv_python == 4.5.5.62
pyenchant == 3.2.2
pytesseract == 0.3.8
yolov5 == 6.0.6
